🌟feat: refactor disruption components for reusability
